htmlpp: Postprocess the output of modern makeinfo.
makeinfo has changed its output for HTML somewhat in the years since this
was written, breaking the postprocessor.  This makes it limp along, working
in a basic way.  The output is readable.  But, really, this should either
use a real HTML parser library or we should abandon it, probably replacing
it by a makeinfo plugin to the extent that it's needed at all.